+++ foo_ui_columns/config_items.cpp
@@ -4,11 +4,13 @@
 #include "status_bar.h"
 
 namespace main_window {
-const char* default_status_pane_script
-    = "// This is the default script for the content of the main status bar pane during "
-      "playback.\r\n\r\n"
-      "$if(%is_status_pane%,%artist% - %title%$crlf(),$if(%ispaused%,Paused,Playing) | )"
-      "%codec% | %bitrate% kbps | %samplerate% Hz | $caps(%channels%) | %playback_time%[ / %length%]";
+
+const char* default_status_bar_script
+    = "// This is the default script for the content of the main status bar section during "
+      "playback.\r\n"
+      "\r\n"
+      "$if(%ispaused%,Paused,Playing) | %codec% | %bitrate% kbps | %samplerate% Hz"
+      " | $caps(%channels%) | %playback_time%[ / %length%]";
 
 const char* default_notification_icon_script
     = "//This is the default script for the content of the notification area icon tooltip "
@@ -19,12 +21,6 @@ const char* default_main_window_title_script
     = "//This is the default script for the title of the main window during playback.\r\n\r\n"
       "[%title% - ]foobar2000";
 
-void on_status_bar_script_change(const char*)
-{
-    if (cui::main_window.get_wnd())
-        SendMessage(cui::main_window.get_wnd(), MSG_UPDATE_STATUS, 0, 0);
-}
-
 void on_main_window_title_script_change(const char*)
 {
     if (cui::main_window.get_wnd())
@@ -32,8 +28,8 @@ void on_main_window_title_script_change(const char*)
 }
 
 fbh::ConfigString config_status_bar_script(
-    GUID{0xb5ca645b, 0xa5e0, 0x4c70, {0xa5, 0x98, 0xcd, 0x62, 0x5c, 0xf3, 0xcc, 0x37}}, default_status_pane_script,
-    &on_status_bar_script_change);
+    GUID{0xb5ca645b, 0xa5e0, 0x4c70, {0xa5, 0x98, 0xcd, 0x62, 0x5c, 0xf3, 0xcc, 0x37}}, default_status_bar_script,
+    [](auto&&) { cui::status_bar::regenerate_text(); });
 
 fbh::ConfigString config_notification_icon_script(
     GUID{0x85d128cf, 0x8b01, 0x4ae9, {0xb8, 0x1c, 0x6b, 0xc4, 0xbe, 0x67, 0x59, 0x9f}},
@@ -59,11 +55,11 @@ cfg_bool cfg_firstrun(GUID{0xf7c0139, 0x8698, 0x42bd, {0xb2, 0xe0, 0x3f, 0x59, 0
         config_get_activate_target_playlist_on_dropped_items_default_value());
 
 cfg_int cfg_transparency_enabled(GUID{0xd5ab3806, 0x8670, 0xf09d, {0x47, 0xdd, 0x1f, 0x45, 0x05, 0x8d, 0x83, 0x8f}},
-    config_get_transparency_enabled_default_value()),
-    cfg_transparency_level(GUID{0x7d50b4ac, 0xfcd4, 0x0a98, {0xd3, 0x64, 0x58, 0xc8, 0x2e, 0x56, 0x19, 0xd7}},
-        config_get_transparency_level_default_value()),
-    cfg_status_show_lock(GUID{0x0ec529ab, 0xa4cb, 0x0fa9, {0xa1, 0xbc, 0x7d, 0x24, 0x56, 0xdc, 0xb5, 0x40}},
-        config_get_status_show_lock_default_value());
+    config_get_transparency_enabled_default_value());
+cfg_int cfg_transparency_level(GUID{0x7d50b4ac, 0xfcd4, 0x0a98, {0xd3, 0x64, 0x58, 0xc8, 0x2e, 0x56, 0x19, 0xd7}},
+    config_get_transparency_level_default_value());
+cfg_int cfg_status_show_lock(GUID{0x0ec529ab, 0xa4cb, 0x0fa9, {0xa1, 0xbc, 0x7d, 0x24, 0x56, 0xdc, 0xb5, 0x40}},
+    config_get_status_show_lock_default_value());
 
 void on_transparency_enabled_change()
 {