+++ foo_ui_columns/main_window.h
@@ -21,7 +21,7 @@ COLORREF get_default_colour(colours::ColourID index, bool themed = false);
 #define ID_PLAYLIST_TOOLTIP 50
 
 /** Main window custom message numbers */
-enum { MSG_SET_AOT = WM_USER + 3, MSG_UPDATE_STATUS, MSG_UPDATE_TITLE, MSG_RUN_INITIAL_SETUP, MSG_NOTICATION_ICON };
+enum { MSG_SET_AOT = WM_USER + 3, MSG_UPDATE_TITLE, MSG_RUN_INITIAL_SETUP, MSG_NOTIFICATION_ICON };
 
 bool remember_window_pos();